+++ lang/id/menu.php
@@ -607,30 +607,3 @@ return [
     'cards' => 'Kartu',
     'charts' => 'Grafik',
 ];
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-