+++ resources/views/partials/menus/_user-account-menu.blade.php
@@ -4,11 +4,11 @@
     $isAltMenu = $menuStyle !== \App\Support\ThemeVersion::default();
     $assetBase = $theme_asset_base ?? \App\Support\ThemeVersion::assetBase($menuStyle);
     $themeVersions = \App\Support\ThemeVersion::available();
-    $profileAvatar =
-        $current_user_display['avatar'] ??
-        asset(($theme_asset_base ?? \App\Support\ThemeVersion::assetBase($menuStyle)) . '/media/avatars/300-1.jpg');
-    $profileName = $current_user_display['name'] ?? 'Guest User';
-    $profileEmail = $current_user_display['email'] ?? '';
+    $authUser = auth()->user();
+    $hasAvatar = !empty($authUser?->avatar_url);
+    $profileName = $authUser?->name ?? ($current_user_display['name'] ?? 'Guest User');
+    $profileEmail = $authUser?->email ?? ($current_user_display['email'] ?? '');
+    $initial = strtoupper(substr($profileName, 0, 1));
 @endphp
 <!--begin::User account menu-->
 <div class="menu menu-sub menu-sub-dropdown menu-column menu-rounded menu-gray-800 menu-state-bg menu-state-color fw-semibold py-4 fs-6 w-275px"
@@ -18,16 +18,22 @@
         <div class="menu-content d-flex align-items-center px-3">
             <!--begin::Avatar-->
             <div class="symbol symbol-50px me-5">
-                <img alt="Logo" src="{{ $profileAvatar }}" />
+                <img alt="Logo" id="header_dropdown_avatar_img" src="{{ $authUser?->avatar_url ?? '' }}" class="w-50px h-50px rounded-3 object-fit-cover {{ $hasAvatar ? '' : 'd-none' }}" />
+                <div id="header_dropdown_avatar_initial" class="symbol-label bg-light-primary text-primary fw-bolder fs-2 rounded-3 w-50px h-50px {{ $hasAvatar ? 'd-none' : '' }}">
+                    {{ $initial }}
+                </div>
             </div>
             <!--end::Avatar-->
             <!--begin::Username-->
             <div class="d-flex flex-column">
                 <div class="fw-bold d-flex align-items-center fs-5">
-                    {{ $profileName }} <span class="badge badge-light-success fw-bold fs-8 px-2 py-1 ms-2">Pro</span>
+                    <span id="header_dropdown_user_name">{{ $profileName }}</span>
+                    @if ($authUser && $authUser->roles->first())
+                        <span class="badge badge-light-success fw-bold fs-8 px-2 py-1 ms-2" id="header_dropdown_user_role">{{ ucfirst($authUser->roles->first()->name) }}</span>
+                    @endif
                 </div>
                 @if ($profileEmail !== '')
-                    <a href="javascript:void(0)" class="fw-semibold text-muted text-hover-primary fs-7">
+                    <a href="javascript:void(0)" class="fw-semibold text-muted text-hover-primary fs-7" id="header_dropdown_user_email">
                         {{ $profileEmail }} </a>
                 @endif
             </div>
@@ -66,8 +72,7 @@
         <div class="menu-sub menu-sub-dropdown w-175px py-4">
             <!--begin::Menu item-->
             <div class="menu-item px-3">
-                <a href="{{ route('pages.account.referrals') }}" class="menu-link px-5"
-                    data-kt-translate="menu.referrals">
+                <a href="{{ route('pages.account.referrals') }}" class="menu-link px-5" data-kt-translate="menu.referrals">
                     {{ __('menu.referrals') }}
                 </a>
             </div>
@@ -81,8 +86,7 @@
             <!--end::Menu item-->
             <!--begin::Menu item-->
             <div class="menu-item px-3">
-                <a href="{{ route('pages.account.statements') }}" class="menu-link px-5"
-                    data-kt-translate="menu.payments">
+                <a href="{{ route('pages.account.statements') }}" class="menu-link px-5" data-kt-translate="menu.payments">
                     {{ __('menu.payments') }}
                 </a>
             </div>
@@ -92,7 +96,7 @@
                 <a href="{{ route('pages.account.statements') }}" class="menu-link d-flex flex-stack px-5">
                     <span data-kt-translate="menu.statements">{{ __('menu.statements') }}</span>
                     <span class="ms-2 lh-0" data-bs-toggle="tooltip"
-                        title="{{ __('menu.view_your_statements') ?? 'View your statements' }}">
+                        title="{{ __('menu.view_your_statements') ?? 'Lihat pernyataan Anda' }}">
                         <i class="ki-duotone ki-information-5 fs-5"><span class="path1"></span><span
                                 class="path2"></span><span class="path3"></span></i> </span>
                 </a>
@@ -130,50 +134,9 @@
     <div class="separator my-2"></div>
     <!--end::Menu separator-->
 
-    {{-- <!--begin::Menu item Language Selection-->
-    <div class="menu-item px-5" data-kt-menu-trigger="{default: 'click', lg: 'hover'}"
-        data-kt-menu-placement="left-start" data-kt-menu-offset="-15px, 0">
-        <a href="javascript:void(0)" class="menu-link px-5">
-            <span class="menu-title position-relative">
-                <span data-kt-translate="menu.language_selection">{{ __('menu.language_selection') }}</span>
-                <span class="fs-8 rounded bg-light px-3 py-2 position-absolute translate-middle-y top-50 end-0">
-                    <span data-kt-element="lang-current-label">{{ \App\Support\LanguageManager::current() === 'id' ? 'Bahasa Indonesia' : 'English' }}</span>
-                    <img class="w-15px h-15px rounded-1 ms-2" data-kt-element="lang-flag-current"
-                        src="{{ asset(($theme_asset_base ?? 'assets') . '/media/flags/' . (\App\Support\LanguageManager::current() === 'id' ? 'indonesia.svg' : 'united-states.svg')) }}" alt="" />
-                </span>
-            </span>
-        </a>
-        <!--begin::Menu sub-->
-        <div class="menu-sub menu-sub-dropdown w-175px py-4">
-            <!--begin::Menu item-->
-            <div class="menu-item px-3">
-                <a href="javascript:void(0)" class="menu-link d-flex px-5 {{ \App\Support\LanguageManager::current() === 'en' ? 'active' : '' }}" data-kt-element="lang-item" data-kt-value="en">
-                    <span class="symbol symbol-20px me-4">
-                        <img class="rounded-1" src="{{ asset(($theme_asset_base ?? 'assets') . '/media/flags/united-states.svg') }}" alt="English" />
-                    </span>
-                    English
-                </a>
-            </div>
-            <!--end::Menu item-->
-            <!--begin::Menu item-->
-            <div class="menu-item px-3">
-                <a href="javascript:void(0)" class="menu-link d-flex px-5 {{ \App\Support\LanguageManager::current() === 'id' ? 'active' : '' }}" data-kt-element="lang-item" data-kt-value="id">
-                    <span class="symbol symbol-20px me-4">
-                        <img class="rounded-1" src="{{ asset(($theme_asset_base ?? 'assets') . '/media/flags/indonesia.svg') }}" alt="Bahasa Indonesia" />
-                    </span>
-                    Bahasa Indonesia
-                </a>
-            </div>
-            <!--end::Menu item-->
-        </div>
-        <!--end::Menu sub-->
-    </div>
-    <!--end::Menu item Language Selection--> --}}
-
     <!--begin::Menu item-->
     <div class="menu-item px-5 my-1">
-        <a href="{{ route('pages.account.settings') }}" class="menu-link px-5"
-            data-kt-translate="menu.account_settings">
+        <a href="{{ route('pages.account.settings') }}" class="menu-link px-5" data-kt-translate="menu.account_settings">
             {{ __('menu.account_settings') }}
         </a>
     </div>